pybackend/requirements.txt
fastapi
uvicorn
python-dotenv
boto3
langchain
langchain-community
langchain-google-genai
faiss-cpu
sentence-transformers
langchain-huggingface
tf-keras
tensorflow
transformers
flask==2.0.1
chromadb 